linux_requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
bottle==0.12.25
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
orjson==3.9.13
packaging==23.2
proxy-tools==0.1.0
pycairo==1.26.0
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyGObject==3.46.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dotenv==1.0.1
pywebview==4.4.1
QtPy==2.4.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.0.post1
